Nicer handling of the initial workspace/ directory

+++ runme.sh
@@ -14,5 +14,8 @@ cp create_keypair.sh ..
 cp clone_repository_as_group.sh ..
 
 # To make it easier: point the default workspace to the checked out version
-ln -s /home/liveuser/IngenieursProject/workspace /home/liveuser/
+if [ ! -e /home/liveuser/workspace ]
+then
+	ln -s /home/liveuser/IngenieursProject/workspace /home/liveuser/
+fi
